refactor(api): move healthcheck to src/shared
