+++ api/src/shared/application/healthcheck/index.js
