+++ api/tests/shared/unit/application/healthcheck/healthcheck-controller_test.js
@@ -1,9 +1,8 @@
-// eslint-disable-next-line @eslint-community/eslint-comments/disable-enable-pair
-/* eslint-disable import/no-restricted-paths */
-import { knex } from '../../../../db/knex-database-connection.js';
-import { healthcheckController } from '../../../../lib/application/healthcheck/healthcheck-controller.js';
-import { redisMonitor } from '../../../../lib/infrastructure/utils/redis-monitor.js';
-import { expect, hFake, sinon } from '../../../test-helper.js';
+// eslint-disable-next-line import/no-restricted-paths
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { redisMonitor } from '../../../../../lib/infrastructure/utils/redis-monitor.js';
+import { healthcheckController } from '../../../../../src/shared/application/healthcheck/healthcheck-controller.js';
+import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | healthcheckController', function () {
   describe('#get', function () {
